pyproject.toml
[project]
name = "startup-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langgraph>=0.6.7",
    "openai>=1.107.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
